+++ app/src/main/java/br/com/vinilapp/core/designsystem/theme/Theme.kt
@@ -5,25 +5,22 @@ import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.darkColorScheme
 import androidx.compose.material3.lightColorScheme
 import androidx.compose.runtime.Composable
-import androidx.compose.ui.graphics.Color
 
 private val LightColors = lightColorScheme(
-    primary = Color(0xFF7B1FA2),
-    secondary = Color(0xFF5D576B),
+    primary = Purple40,
+    secondary = Neutral40
 )
 
 private val DarkColors = darkColorScheme(
-    primary = Color(0xFFE1B6FF),
-    secondary = Color(0xFFCCC2DB),
+    primary = Purple80,
+    secondary = Neutral80
 )
 
 @Composable
-fun VinilAppTheme(
-    darkTheme: Boolean = isSystemInDarkTheme(),
-    content: @Composable () -> Unit,
-) {
+fun VinilAppTheme(darkTheme: Boolean = isSystemInDarkTheme(), content: @Composable () -> Unit) {
     MaterialTheme(
         colorScheme = if (darkTheme) DarkColors else LightColors,
-        content = content,
+        typography = VinilTypography,
+        content = content
     )
 }