+++ app/src/main/java/br/com/vinilapp/data/notification/NotificationDataSource.kt
@@ -1,0 +1,4 @@
+package br.com.vinilapp.data.notification
+
+/** Contrato futuro para leitura de notificações de mídia autorizadas pelo usuário. */
+interface NotificationDataSource