+++ app/src/main/java/br/com/vinilapp/feature/nowplaying/presentation/NowPlayingRoute.kt
@@ -1,11 +1,11 @@
 package br.com.vinilapp.feature.nowplaying.presentation
 
 import androidx.compose.runtime.Composable
+import androidx.hilt.navigation.compose.hiltViewModel
 import androidx.lifecycle.compose.collectAsStateWithLifecycle
-import androidx.lifecycle.viewmodel.compose.viewModel
 
 @Composable
-fun NowPlayingRoute(viewModel: NowPlayingViewModel = viewModel()) {
+fun NowPlayingRoute(viewModel: NowPlayingViewModel = hiltViewModel()) {
     viewModel.uiState.collectAsStateWithLifecycle()
     NowPlayingScreen()
 }